Fix countdown display skipping seconds This is an attempt to fix the second-skipping display bug.
Instead of changing the tick display based on the "wall time" this
decrements the seconds counter based on the timer interval itself,
which should be the correct behavior since it won't be affected by
alignment with the system clock.

An additional check was added to deal with cases when the thread
is paused e.g. when the lid is closed on a laptop or the user is
interacting with the menu bar.